Move conversion and more explicit error for UserDistributionID (#55)
