+++ crates/wslplugins-rs/src/distribution_id.rs
@@ -1,25 +1,24 @@
-//! # Module `DistributionID`
+//! Distribution identifiers used by the crate.
 //!
-//! This module defines an abstraction to represent WSL distributions through a
-//! [`DistributionID`]. It supports two types of identifiers: system-level distributions
-//! and user-specific installed distributions identified by a GUID.
+//! [`DistributionID`] models the two kinds of distributions exposed by WSL:
+//! the shared system distribution and user-installed distributions identified by
+//! a [`UserDistributionID`].
 //!
-//! ## Key Features
+//! The module also exposes the conversions commonly needed by the API surface:
 //!
-//! - Bi-directional conversion between [`DistributionID`] and [`UserDistributionID`].
-//! - Robust error handling for conversions via [`ConversionError`].
-//! - Display implementation ([Display]) and support for other idiomatic conversions.
+//! - converting from a [`UserDistributionID`] or `Option<UserDistributionID>`
+//!   into a [`DistributionID`],
+//! - converting a [`DistributionID`] back into `Option<UserDistributionID>`,
+//! - retrieving a distribution identifier from any
+//!   [`CoreDistributionInformation`] implementation.
 //!
-//! ## Usage Context
-//!
-//! This abstraction is particularly useful in environments where WSL requires
-//! distribution identification via GUIDs or when a distinction between a system-level
-//! distribution and a user-specific distribution is necessary. The associated functions
-//! and conversions simplify integration with APIs like those defined in `WslPluginApi`.
+//! When a caller needs a user distribution identifier and receives
+//! [`DistributionID::System`] instead, [`UserDistributionIDConversionError`] is returned.
 
 use crate::{CoreDistributionInformation, UserDistributionID};
-use std::{convert::TryFrom, fmt::Display};
-use thiserror::Error;
+use std::fmt::Display;
+
+pub use crate::user_distribution_id::UserDistributionIDConversionError;
 
 /// Represents a distribution identifier in the Windows Subsystem for Linux (WSL).
 ///
@@ -62,22 +61,6 @@ impl DistributionID {
     #[inline]
     pub const fn is_system(&self) -> bool {
         matches!(*self, Self::System)
-    }
-}
-
-/// Error type for conversion failures between [`DistributionID`] and [`UserDistributionID`].
-#[derive(Debug, Error, Clone, Copy, PartialEq, Eq, Hash)]
-#[error("Cannot convert System distribution to UserDistribution.")]
-pub struct ConversionError;
-
-impl TryFrom<DistributionID> for UserDistributionID {
-    type Error = ConversionError;
-    #[inline]
-    fn try_from(value: DistributionID) -> Result<Self, Self::Error> {
-        match value {
-            DistributionID::User(id) => Ok(id),
-            DistributionID::System => Err(ConversionError),
-        }
     }
 }
 
@@ -177,14 +160,6 @@ mod tests {
     #[test]
     fn is_user_returns_false_for_system_distribution() {
         assert!(!DistributionID::System.is_user());
-    }
-
-    #[test]
-    fn try_from_system_returns_error() {
-        assert_eq!(
-            UserDistributionID::try_from(DistributionID::System),
-            Err(ConversionError)
-        );
     }
 
     #[test]
