+++ crates/wslplugins-rs/src/user_distribution_id.rs
@@ -2,18 +2,24 @@ use std::{
     fmt::{Debug, Display, LowerHex, UpperHex},
     str::FromStr,
 };
+use thiserror::Error;
 pub mod fmt;
 mod parse_error;
 use fmt::DefaultFormatter;
 pub use parse_error::ParseError;
 
-use crate::CoreDistributionInformation;
+use crate::{CoreDistributionInformation, DistributionID};
 #[cfg(feature = "uuid")]
 mod uuid_impl;
 
 #[repr(transparent)]
 #[derive(Clone, Copy, Default, PartialEq, Eq, Hash)]
 pub struct UserDistributionID(pub windows_core::GUID);
+
+/// Error type for conversion failures between [`DistributionID`] and [`UserDistributionID`].
+#[derive(Debug, Error, Clone, Copy, PartialEq, Eq, Hash)]
+#[error("Cannot convert System distribution to UserDistribution.")]
+pub struct UserDistributionIDConversionError;
 
 impl From<windows_core::GUID> for UserDistributionID {
     #[inline]
@@ -40,6 +46,17 @@ impl<T: CoreDistributionInformation> From<&T> for UserDistributionID {
     #[inline]
     fn from(value: &T) -> Self {
         value.id()
+    }
+}
+
+impl TryFrom<DistributionID> for UserDistributionID {
+    type Error = UserDistributionIDConversionError;
+    #[inline]
+    fn try_from(value: DistributionID) -> Result<Self, Self::Error> {
+        match value {
+            DistributionID::User(id) => Ok(id),
+            DistributionID::System => Err(UserDistributionIDConversionError),
+        }
     }
 }
 
@@ -115,5 +132,13 @@ mod tests {
             prop_assert_eq!(format!("{:X}", user_dist_id), format!("{:?}", user_dist_id));
             prop_assert_eq!(format!("{:x}", user_dist_id), format!("{user_dist_id:X}").to_ascii_lowercase());
         }
+    }
+
+    #[test]
+    fn try_from_system_returns_error() {
+        assert_eq!(
+            UserDistributionID::try_from(DistributionID::System),
+            Err(UserDistributionIDConversionError)
+        );
     }
 }
